Auto-Forward-Bot-V2/requirements.txt
motor
Dnspython
cryptg
tgcrypto
Pyrogram==2.0.93
psutil
opencv-python-headless
python-decouple
requests
speedtest-cli
pymongo
aiohttp
